+++ packages/mobile_sidebar/test/mobile_sidebar_test.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:mobile_sidebar/mobile_sidebar.dart';
+
+void main() {
+  const MethodChannel channel = MethodChannel('mobile_sidebar');
+
+  setUp(() {
+    channel.setMockMethodCallHandler((MethodCall methodCall) async {
+      return '42';
+    });
+  });
+
+  tearDown(() {
+    channel.setMockMethodCallHandler(null);
+  });
+
+ 
+}