+++ packages/persist_theme/lib/persist_theme.dart
@@ -1,3 +1,2 @@
-export 'data/classes/theme.dart';
 export 'data/models/theme_model.dart';
 export 'ui/theme_widgets.dart';