+++ packages/flutter_whatsnew/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.0 - 04.06.2019
+
+* Making Example Desktop Aware
+
+## 0.2.1 - 04.05.2019
+
+* `tb2761` Added Button Colors
+
 ## [0.2.0] - Breaking Changes
 
 * Updated Example